+++ config/routes.rb
@@ -123,6 +123,7 @@ Rails.application.routes.draw do
       post 'list', action: 'list', on: :collection
       put 'cancel', on: :member
       get 'download', on: :member
+      get 'items/:id', action: 'show_item', on: :collection
       post 'items/:id/cash_check', action: 'cash_check', on: :collection
       post 'items/:id/refresh_item', action: 'refresh_item', on: :collection
       post 'items/:id/pay_item', action: 'pay_item', on: :collection